requirements.txt
tflite-runtime==2.14.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart
pillow==10.3.0
numpy==1.26.4